hopefully fixing the crashes that Simon reported: * the progress dialog must not close before the background   thread actually terminated. * remove the debug-hack
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@19919 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/SVNProgressDlg.cpp
@@ -1270,17 +1270,18 @@ void CSVNProgressDlg::OnOK()
 LRESULT CSVNProgressDlg::OnCloseOnEnd(WPARAM /*wParam*/, LPARAM /*lParam*/)
 {
     m_bCancelled = TRUE;
-    if (m_bThreadRunning)
-    {
-        if (WaitForSingleObject(m_pThread->m_hThread, 1000) != WAIT_OBJECT_0)
-        {
-            // end the process the hard way
-            TerminateProcess(GetCurrentProcess(), 0);
-        }
-    }
-    else
-        __super::OnOK();
 
+    // always wait for the thread to truely terminate.
+    // (this function can only be triggered from that thread)
+
+    if (   (m_pThread != NULL) 
+        && (WaitForSingleObject(m_pThread->m_hThread, 1000) != WAIT_OBJECT_0))
+    {
+        // end the process the hard way
+        TerminateProcess(GetCurrentProcess(), 0);
+    }
+
+    __super::OnOK();
     return 0;
 }
